refactor: 관리자 예약 목록 리스트 조회 api 엔드포인트 리소스를 예약으로 변경
