+++ src/main/resources/static/js/waiting.js
@@ -1,6 +1,6 @@
 document.addEventListener('DOMContentLoaded', () => {
 
-    fetch('/api/admin/waitings') // 내 예약 목록 조회 API 호출
+    fetch('/api/admin/reservations/waiting-list') // 내 예약 목록 조회 API 호출
         .then(response => {
             if (response.status === 200) return response.json();
             throw new Error('Read failed');
