--- src/test/java/roomescape/controller/api/WaitingApiControllerTest.java
@@ -1,38 +1,0 @@
-package roomescape.controller.api;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
-import roomescape.util.TokenGenerator;
-
-import static org.hamcrest.Matchers.is;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class WaitingApiControllerTest {
-
-    @Test
-    @DisplayName("모든 예약 대기 목록을 가져온다.")
-    void selectWaitings() {
-        RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .cookie("token", TokenGenerator.makeAdminToken())
-                .when().get("/api/admin/waitings")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(4));
-    }
-
-    @Test
-    @DisplayName("예약 대기 삭제를 정상적으로 수행한다.")
-    void deleteWaiting() {
-        RestAssured.given().log().all()
-                .cookie("token", TokenGenerator.makeAdminToken())
-                .when().delete("/api/reservations/2")
-                .then().log().all()
-                .statusCode(204);
-    }
-}
